admins can delete any comment & recipe

+++ Smakoowa-Api/Smakoowa-Api/Services/RecipeService.cs
@@ -44,6 +44,9 @@
             var recipe = await _recipeRepository.FindByConditionsFirstOrDefault(c => c.Id == recipeId);
             if (recipe == null) return ServiceResponse.Error($"Recipe with id: {recipeId} not found.");
 
+            if (recipe.CreatorId != _apiUserService.GetCurrentUserId() && !_apiUserService.UserIsAdmin())
+                return ServiceResponse.Error($"User isn't the owner of recipe with id: {recipeId}.");
+
             try
             {
                 await _recipeRepository.Delete(recipe);
